+++ src/main/java/ladder/domain/direction/Direction.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.direction;
 
 import java.util.Arrays;
