+++ src/main/java/ladder/domain/user/User.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.user;
 
 import java.util.regex.Pattern;
