+++ src/test/java/ladder/domain/ladder/LineGeneratorTest.java
@@ -1,5 +1,11 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
+import ladder.domain.direction.Direction;
+import ladder.domain.direction.DirectionGeneratorImpl;
+import ladder.domain.direction.DirectionNeutralGeneratorImpl;
+import ladder.domain.direction.DirectionRightGeneratorImpl;
+import ladder.domain.ladder.Line;
+import ladder.domain.ladder.LineGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
