+++ src/test/java/ladder/domain/user/UserTest.java
@@ -1,7 +1,8 @@
-package ladder.domain;
+package ladder.domain.user;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import ladder.domain.user.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
