+++ src/test/java/ladder/domain/user/UsersTest.java
@@ -1,8 +1,10 @@
-package ladder.domain;
+package ladder.domain.user;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
+import ladder.domain.user.User;
+import ladder.domain.user.Users;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
